[UE-5] Initial skeleton for wrapper package
The intention for this change is to add the necessary packages
needed for the wrapper as well as testing the wrapper class.
For the purposes of the PR, a simple method returning a boolean
is added to the wrapper class. The test tests this method to make
sure that the package is set up to correctly use Jest.

ps-id: 44D3D9A4-57CE-44DC-8625-97D5BEAC86EE