[UE-5] Init and isOn functions for wrapper
The intention for this change is to add the init and initForTest
functions that make this wrapper functional, along with a helper
method to create the Growthbook client instance. Since the test init
function is not actually making a call to the api, the fetch function
is omitted from the method. The seedsToGBFeatures method formats
the passed in seeds to match with the types returned from Growthbook
when we fetch from the API.  The isOn method is also added to
give important functionality as well as an easy method to use in
testing as more features are added.

ps-id: F9AB1E3E-C8C8-41D0-AD2F-F55312FC2173