[UE-24] Document initializing Typescript Toggles
The intention for this change is to add documentation for setting
up a user's app to use the wrapper. This includes creating the
singleton that holds the user's instance of the wrapper and
firing the init function in the same place where the app is initialized.
In the example, an Express app is used as that is a common framework
to be used with Node.

It also includes examples of using the `isOn` method within your app as
well as using the `initForTests` method in the user's unit tests

ps-id: 545094BB-EA82-4791-BAEF-D1F3BF252145