[UE-5] Add refresh toggle method
The intention for this change is to add a specific method
for fetching the latest features from Growthbook and adding
them to the wrapper's instance. Since the same work was
already being done in the init function, that logic was moved
to a separate public method. It also handles if the API fetch
is unsuccessful by logging the error from Growthbook.

ps-id: 0FEDA8EA-3570-430C-9D30-6C564C007E75