[UE-5] Add overrides parameter to init
The intention for this change is allow users to add overrides
to the Growthbook features fetched at init. It is added as an
optional parameter to both init functions and stored to use in
the isOn method as a conditional before checking Growthbook
features.

The parameters are also changed in the init functions to have
to use parameter names in the function call. This allows for
overrides to be added as a parameter without having to add
seeds. It will also be helpful in the future when attributes are
added as a parameter to init.

[changelog]
added: overrides parameter to init functions

ps-id: D408DA4B-ADD7-4415-8E7B-2DB32185474C