[UE-5] Document how to get overrides from env
The intention for this change is to clearly explain to users
both methods of adding overrides to their wrapper. It gives
an example of how to add the overrides parameter to the
init function, as well as explaining how to add overrides to
env variables with an example of the acceptable formatting
for env variables to be correctly read by the get override
key method.

[changelog]
added: 'Adding Overrides' section to README

ps-id: 8013A1B2-1A36-4A5C-A230-1F9962407331