[UE-5] Add rules parameter to test init
The intention for this change is to be able to add a list of
rules to the test init function. These are needed to correctly
mock the features with override values that we get from
Growthbook. This is the first step needed to set attributes
on and after init.

For example, if the following rules argument is passed to the test
init function:
```[
    {
      'condition': {
        'version': {'$gt': '1.0.0'}
      },
      'force': true
    }
];```
The test client will be able to check for a 'version' attribute and
toggle the feature on or off based on the attribute's value (will
toggle on if the value is greater than 1.0.0.

[changelog]
added: rules parameter to test init function

<!-- ps-id: 43ed9329-b74c-46f1-8cf8-1364b1f2193f -->