[UE-5] Add attributes parameter to init
The intention for this change is to be able to add attributes
from the app (ie version number) to the init functions. This
allows for version baselining or checking any other attributes
that can be obtained on initialization of the app.

The attributes are saved to a value on the class and used in the
create client method when creating the instance of Growthbook.
Tests for this  functionality use a version number attribute to test
against rules that dictate the app should be a certain version or
higher to use the mock feature.

[changelog]
added: attributes parameter to init

<!-- ps-id: e48509ea-b603-483c-be2a-0df4a8ba223b -->