[UE-5] Add public set attribute method
The intention for this change is to give users the ability to add
attributes after the wrapper has been initialized. This is helpful
in instances where required attributes cannot be acquired until
long after an app is initialized. For instance, with canary testing
a feature will only be available to select users. The user's id
cannot be determined until after they have logged in.

To achieve this, a new public method has been added to the
wrapper that takes in an attributes parameter. If the client
already has attributes (ie they were added during init), the
new attributes are added alongside the old attributes.
Otherwise, the new attributes are assigned to the client
using the existing `setAttributes` method from the sdk.

The tests tests all possible permutations of a different version
number to ensure version comparison is working properly in
node.

[changelog]
added: public set attribute method

<!-- ps-id: 17ee3ac6-d2dc-45f8-81c1-131ee71ad296 -->