[UE-39] Fetch non-boolean toggle state
The reason for this change is to give users of this package
a public method to get the value of a feature. Previously,
they could only retrieve whether a feature was on or off.
This allows them to use the non boolean value in their
applications

[changelog]
added: get value method in wrapper

<!-- ps-id: d8707451-6c50-405c-ad24-c983f685dad8 -->