[UE-26] Document non-boolean toggle state
The reason for this change is to give users of this package
instruction on how to use the value method. The example
shows how they might get a feature's value (in this case
an int) and use it to determine some business logic.

[changelog]
added: 'Evaluate Feature value' section in README

<!-- ps-id: 4fb8f63e-8ec4-43cc-87ee-dbbdc45047c4 -->