Support GrowthBook SDK Connector style
Support GrowthBook SDK Connector style so that it would be clearer how
to initialize the SDK and be in alignment with their updated
documentation.

As part of this I also added in an full integration e2e test suite so
that we could test the library against the actual service when we want
to. Also I think this will give better examples for people in terms of
actual usage.

Beyond that I changed the constructor to take the apiHost and clientKey.
I also upgraded to the latest version of the javascript sdk.

[changelog]
changed: construction to take apiHost and clientKey

skip_ticket_id_check

<!-- ps-id: 953fa8e6-514a-4319-bd29-879e8001a81d -->