Update README with install instructions
The README did not include install instructions for the package.  As a developer, this is probably the first thing I would look for, which happened when I was trying to install it and had to ask a fellow developer for the install path.  So, this should be added to the README to avoid developer confusion and aide the install flow.

Update README with instructions on how to install the SDK.