Update growthbook sdk to v1
In order to be on a real release instead of a pre-release build, this upgrades growthbook to the latest and greatest. It appears there are no breaking changes except that the `loadFeatures` method has been deprecated in favor of `init`, which I have swapped out here.

[changelog]
changed: updated underlying growthbook package to v1.6.1